requirements.txt
gradio==4.2.0
openai==1.2.3
python-dotenv==1.0.0
typing_extensions==4.8.0
